% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle-nbs.R
\name{adjacent_cycles}
\alias{adjacent_cycles}
\title{Construct adjacency matrix of neighbourhood cycles}
\usage{
adjacent_cycles(cycles, measure = "median")
}
\arguments{
\item{cycles}{List of cycles obtained from \link{network_cycles}.}

\item{measure}{Measure of aggregate centrality; acceptable values are
"median", "mean", "max", "min".}
}
\value{
A `data.frame` of three columns:
\enumerate{
\item from - cycle from which connection is made
\item to - cycle to which connection is made
\item centrality - Median centrality of edges connecting those two cycles
}
}
\description{
Construct adjacency matrix of neighbourhood cycles
}
