% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-nbs.R
\name{cut_nbs}
\alias{cut_nbs}
\title{Use result of \link{neighbourhoods} function to make and score an LTN}
\usage{
cut_nbs(nbs, i)
}
\arguments{
\item{nbs}{Results of \link{neighbourhoods} function.}

\item{i}{Index of which neighbour pair is to be cut.}
}
\description{
Use result of \link{neighbourhoods} function to make and score an LTN
}
