% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTN-package.R
\docType{package}
\name{LTN-package}
\alias{LTN}
\alias{LTN-package}
\title{LTN: Planning tool for Low Traffic Neighbourhoods}
\description{
Planning tool for Low Traffic Neighbourhoods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ATFutures-labs/LTN}
  \item Report bugs at \url{https://github.com/ATFutures-labs/LTN/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
