% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltn.R
\name{ltn}
\alias{ltn}
\title{ltn}
\usage{
ltn(net)
}
\arguments{
\item{net}{Street network to be analysed}
}
\value{
List of sub-networks, each of which represents a candidate LTN
}
\description{
Convert a street network into a list of Low Traffic Neighbourhoods (LTNs).
}
