% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltn-score.R
\name{ltn_score}
\alias{ltn_score}
\title{Score an LTN formed by blocking one street segment between two adjacent
neighbourhoods.}
\usage{
ltn_score(nbs, index, dmax)
}
\arguments{
\item{nbs}{Output of main \link{neighbourhoods} function.}

\item{index}{Index into rows of `nbs$nbs` specifying which pairs of adjacent
neighbourhoods are to be scored.}

\item{dmax}{Maximal distance in metres around neighbourhood to use to
generate scores.}
}
\value{
Modified version of `nbs$nbs` from input parameter, reduced to only
those neighbour pairs specified in `index`, and with additional column,
`pop_decr_in` and `pop_incr_out` specifying absolute decreases within
and increases surrounding proposed LTN.
}
\description{
Score an LTN formed by blocking one street segment between two adjacent
neighbourhoods.
}
