% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods-package.R
\docType{package}
\name{neighbourhoods-package}
\alias{neighbourhoods}
\alias{neighbourhoods-package}
\title{neighbourhoods: Efficient Identification of Neighbourhoods within Networks}
\description{
Algorithm for efficient identification of neighbourhoods within
    networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ATFutures-labs/neighbourhoods}
  \item Report bugs at \url{https://github.com/ATFutures-labs/neighbourhoods/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
