% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{neighbourhoods}
\alias{neighbourhoods}
\title{Find candidate low-traffic-neighbourhoods in an input street network.}
\usage{
neighbourhoods(network, popdens)
}
\arguments{
\item{network}{Street network in \pkg{silicate} `sc` format, extracted with
\pkg{dodgr} function, `dodgr_streetnet_sc`.}

\item{popdens}{Path to local population density file covering region of street
network, and in `geotiff` format.}
}
\value{
A `data.frame` of candidate low-traffic neighbourhoods.
}
\description{
Find candidate low-traffic-neighbourhoods in an input street network.
}
