% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{network_cycles}
\alias{network_cycles}
\title{network_cycles}
\usage{
network_cycles(x)
}
\arguments{
\item{x}{An \pkg{dodgr} street network processed with the
`dodgr_contract_graph` and `merge_directed_graph` functions.}
}
\value{
A list of the minimal cycles of the street network, each of which has
three columns of (`.vx0`, `.vx1`, `.edge_`).
}
\description{
Get the minimal cycles of an undirected version of a \pkg{dodgr} street
network in contracted and undirected form.
}
