% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncontract-cycles.R
\name{uncontract_cycles}
\alias{uncontract_cycles}
\title{Convert cycles created on contracted graph back to equivalent uncontracted
cycles.}
\usage{
uncontract_cycles(paths, graph)
}
\arguments{
\item{paths}{List of cycle paths as a result of \link{network_cycles}.}

\item{graph}{Contracted graph resulting from call to `dodgr_contract_graph`.}
}
\value{
Equivalent list of `paths`, with each path expanded out to full
edges in original, non-contracted graph.
}
\description{
Convert cycles created on contracted graph back to equivalent uncontracted
cycles.
}
