% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle-nbs.R
\name{adjacent_cycles}
\alias{adjacent_cycles}
\title{Construct adjacency matrix of neighbourhood cycles}
\usage{
adjacent_cycles(cycles)
}
\arguments{
\item{cycles}{List of cycles obtained from \link{network_cycles}.}
}
\value{
A `data.frame` of three columns:
\enumerate{
\item from - cycle from which connection is made
\item to - cycle to which connection is made
\item edges - List-column of all shared edges between (from, to) pair.
}
}
\description{
Construct adjacency matrix of neighbourhood cycles
}
