% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods-package.R
\docType{data}
\name{hampi_sc}
\alias{hampi_sc}
\title{hampi_sc}
\format{
A \pkg{silicate} representation of the way network of Hampi.
}
\description{
A sample street network from the primarily pedestrian township of Hampi,
Karnataka, India.
}
\keyword{datasets}
