% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltn.R
\name{ltn}
\alias{ltn}
\title{ltn}
\usage{
ltn(net, verts)
}
\arguments{
\item{net}{Street network to be analysed}

\item{verts}{Result of call to \link{dodgr_centrality} with `edges = FALSE`
and so returning vertex-based centrality.}
}
\value{
List of sub-networks, each of which represents a candidate LTN
}
\description{
Convert a street network into a list of Low Traffic Neighbourhoods (LTNs).
}
